***** DO NOT USE IF YOU NEED TrendType.Daily or TrendType.Weekly THEY ARE NOT WORKING ****
Part two of the transition to pure JSON for the Trend endpoints. This pass :
Removes CountryType from Location (as this was ALWAYS "Country" in the Xml, and never supplied for the Json).
Adds marker interface IRequestProcessorWantsJson that xxxRequestProcessor classes can "implement" to indicate they want pure unchanged Json responses instead of the munged-to-Xml responses.
Added GetDate helper for string values.
Added Deserialize and DeserializeArray for use by DataContractJsonSerializer objects.
TrendRequestProcessor now does everything in pure Json format. BROKEN FOR NOW AS Daily and Weekly don't do anything.
Adds some extra debugging support in TwitterExecute so we get the non-Xml, non-Json response bodies in the error status (lets us see the whale page).
Update the LinqToTwitterDemo/Program to allow optional use of various authentication without recompile.  Switched TrendsDemos to use non-worldwide for the TrendType.Location so we're testing something different than TrendType.Trend (which is an alias now of location woeid == 1)